relocated base child stream class
