+++ tap_ms_graph/client.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import json
 import uuid
-
-# from pathlib import Path
 from typing import Any, Dict, Generator, Iterable, Optional, Union
 from urllib.parse import parse_qsl, urljoin
 
@@ -176,3 +174,29 @@ class MSGraphStream(RESTStream):
 
     def get_child_context(self, record: dict, context: Optional[dict]) -> dict:
         return {v: record[k] for k, v in self.child_context.items()}
+
+
+class MSGraphChildStream(MSGraphStream):
+    parent_context_schema: dict = {}
+
+    @property
+    def schema(self):
+        schema = super().schema
+        odata_type = {"@odata.type": {"type": ["string", "null"]}}
+
+        schema["properties"] = {
+            **self.parent_context_schema,
+            **odata_type,
+            **schema["properties"],
+        }
+
+        return schema
+
+    def post_process(self, row: dict, context: dict | None = None) -> dict:
+        processed_row = super().post_process(row, context) or {}
+
+        if context:
+            context_fields = {k: context.get(k, "") for k in context.keys()}
+            return {**context_fields, **processed_row}
+
+        return processed_row
